Moved expressions parsing to exprVisitor
